// FinalProject/TrafficOffenceManagementSystem/TrafficOffenceManagementSystemFrontend/src/app/model/officer.ts
export class Officer {

    policeId: number;
    name: string;
    address: string;
    nic: string;
    phone: number;
    gender: string;
    dob: Date;
    password: string;
    email: string;
/*
    constructor(policeId: number, name: string, address: string, nic: string, phone: number, gender: string, dob: Date, password: string){

        this.policeId = policeId;
        this.name = name;
        this.address = address;
        this.nic = nic;
        this.phone = phone;
        this.gender = gender;
        this.dob = dob;
        this.password = password;
    }*/

}