// FinalProject/TrafficOffenceManagementSystem/TrafficOffenceManagementSystemFrontend/src/app/model/user.ts
export class User {

     username: string;
     password: string;
     email: string;
     enabled: boolean;
     accountNonExpired: boolean;
     credentialsNonExpired: boolean;
     accountNonLocked: boolean;

 /*   public constructor(name, nic, contact, email, password){
        this.name = name;
        this.nic = nic;
        this.contact = contact;
        this.email = email;
        this.password = password;
    }
    */
}